#23 トップページのコンポーネントの分割
