+++ app/layout.tsx
@@ -1,11 +1,7 @@
-import { EnvVarWarning } from "@/components/env-var-warning";
-import HeaderAuth from "@/components/header-auth";
-import { hasEnvVars } from "@/utils/supabase/check-env-vars";
+import Navbar from "@/components/navbar";
 import { ThemeProvider } from "next-themes";
 import { Geist } from "next/font/google";
-import Link from "next/link";
 import "./globals.css";
-import Image from "next/image";
 
 const defaultUrl = process.env.VERCEL_URL
   ? `https://${process.env.VERCEL_URL}`
@@ -38,21 +34,7 @@ export default function RootLayout({
         >
           <main className="min-h-screen flex flex-col items-center">
             <div className="flex-1 w-full flex flex-col items-center">
-              <nav className="w-full flex justify-center border-b border-b-foreground/10 h-16">
-                <div className="w-full flex justify-between items-center px-4 lg:px-6 text-sm">
-                  <div className="flex gap-5 items-center font-semibold">
-                    <Link href="/">
-                      <Image
-                        src="/img/logo.png"
-                        alt="logo"
-                        width={57}
-                        height={48}
-                      />
-                    </Link>
-                  </div>
-                  <HeaderAuth />
-                </div>
-              </nav>
+              <Navbar />
               <div className="flex flex-col py-4">{children}</div>
 
               {/* <footer className="w-full flex items-center justify-center border-t mx-auto text-center text-xs gap-8 py-16">
